Tests/LibWeb/Text/expected/Streams/init-from-cloned-fetch-response.txt

Was able to create a reader from a cloned Fetch response!
